Fixed a seeking bug;
I shot myself int he foot by adding the expected_pageno arg to
stream_reset; it needed the out of sequence trip to skip past
continued packets.

Removed the argument; stream_reset no sets pageno to -1 (not 0 like in
init) so that seeks are flagged; the out of sequence handling trips,
but the -1 surpresses the out of sequence error.

Monty


git-svn-id: http://svn.xiph.org/trunk/vorbis@301 0101bb08-14d6-0310-b084-bc0e0c8e3800