Trac 2139 indirectly brought to light the case of a stream that uses a single-entry codebook, but does not code a codeword of length 1 equal to zero.  Such a stream could cause a stream to read garbage.
There is no apparent chance of garbage memory writes as this happen
entirely after decode setup, however there is playback DoS potential.

This commit special cases single-entry codebook setup so that decode
is well-defined for streams with single-entry codebooks, and adds some
comments to make it more clear how the case is handled.



git-svn-id: http://svn.xiph.org/trunk/vorbis@19444 0101bb08-14d6-0310-b084-bc0e0c8e3800